Extract terminal quick command discovery
